pyproject.toml
[project]
name = "uptimerobot-mcp"
version = "1.0.0"
description = "MCP server for UptimeRobot monitoring API v2"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "84EM", email = "hello@84em.com" }]
keywords = ["mcp", "uptimerobot", "monitoring", "uptime", "model-context-protocol"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "fastmcp>=3.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.1.0",
]

[project.urls]
Homepage = "https://github.com/84em/uptimerobot-mcp"
Repository = "https://github.com/84em/uptimerobot-mcp"
Issues = "https://github.com/84em/uptimerobot-mcp/issues"
Changelog = "https://github.com/84em/uptimerobot-mcp/blob/main/CHANGELOG.md"

[project.scripts]
uptimerobot-mcp = "uptimerobot_mcp.server:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["uptimerobot_mcp*"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9.0",
    "mypy>=1.14",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
